Edit form, sorting and usability improvements
Edit form works now. Just a few interface tweaks needed before
deployment.
Object sorting added to model. This enables specifying the sort order
for lists of sqlalchemy objects in the model itself.
Some javascript usability improvements in the webinterface related
to the dialog scrolling out of focus, and scrolling to the bottom of a
page.